add systick part

+++ firmware/Universal_FW/User/code/rgbled.h
@@ -37,9 +37,9 @@ struct SelPin {
 };
 
 typedef struct Rgb {
-    uint8_t r;
-    uint8_t g;
-    uint8_t b;
+    uint16_t r;
+    uint16_t g;
+    uint16_t b;
 } Rgb;
